fix: refactor code, shift handling reject call logic from jetpack compose to activity
